Correctly pass Copybara config to container (#1)
* Correctly pass Copybara config to container

Copybara Docker images uses a wrapper script instead of directly passing
arguments to Copybara itself, in order to set arguments for the JVM.
The script is receiving its arguments through environment variable,
instead of commandline args.

* Ignore copybara return code

Copybara exits with error code 4 when the workflow would produce no
changes. This means the build fails, so ignore the error code for now.